+++ Gameboy/Gameboi/Hardware/IUpdateable.cs
@@ -2,6 +2,6 @@ namespace GB_Emulator.Gameboi.Hardware
 {
     public interface IUpdatable
     {
-        void Update(byte cycles, ulong speed);
+        void Update(uint cycles, ulong speed);
     }
 }
